# Conflicts:
#	backend/team6/src/main/java/programmers/team6/domain/vacation/entity/VacationInfo.java
#	backend/team6/src/main/java/programmers/team6/domain/vacation/repository/JdbcVacationEligibilitiesRepository.java
#	backend/team6/src/main/java/programmers/team6/domain/vacation/repository/VacationInfoRepository.java
#	backend/team6/src/main/java/programmers/team6/domain/vacation/rule/AnnualVacationGrantRule.java
#	backend/team6/src/main/java/programmers/team6/domain/vacation/rule/AnnualVacationRule.java
#	backend/team6/src/main/java/programmers/team6/domain/vacation/rule/vacationgranteligiblities/VacationGrantEligibility.java
#	backend/team6/src/main/java/programmers/team6/domain/vacation/rule/vacationgranteligiblities/VacationGrantInfo.java
#	backend/team6/src/test/java/programmers/team6/domain/vacation/entity/VacationInfoTest.java
#	backend/team6/src/test/java/programmers/team6/domain/vacation/repository/JdbcVacationEligibilitiesRepositoryTest.java
#	backend/team6/src/test/java/programmers/team6/domain/vacation/service/VacationInfoServiceTest.java

+++ backend/team6/src/test/java/programmers/team6/domain/vacation/entity/VacationInfoTest.java
@@ -11,7 +11,7 @@ class VacationInfoTest {
 	@Test
 	void 휴가정보업데이트() {
 		VacationInfo info = new VacationInfo(15, 13, "test", 1L);
-		int updateTotalCount = 13;
+		double updateTotalCount = 13;
 
 		VacationInfoUpdateResult result = info.updateTotalCount(updateTotalCount);
 
@@ -21,9 +21,9 @@ class VacationInfoTest {
 
 	@Test
 	void 사용휴가정보보다_적게_부여휴가를_비업데이트() {
-		int totalCount = 15;
+		double totalCount = 15;
 		VacationInfo info = new VacationInfo(totalCount, 13, "test", 1L);
-		int updateTotalCount = 12;
+		double updateTotalCount = 12;
 
 		VacationInfoUpdateResult result = info.updateTotalCount(updateTotalCount);
 
@@ -31,4 +31,5 @@ class VacationInfoTest {
 		assertThat(result).isEqualTo(VacationInfoUpdateResult.MISS_RULES);
 		assertThat(totalCount).isEqualTo(info.getTotalCount());
 	}
+
 }
